# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "mementos-backend",
        "description": "Documentation de l'API du projet Mementos",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/cartes/{themeId}/user": {
            "get": {
                "tags": [
                    "Cartes"
                ],
                "summary": "Récupérer les cartes d'un thème pour l'utilisateur connecté",
                "description": "Retourne toutes les cartes associées à un thème spécifique appartenant à l'utilisateur connecté.",
                "operationId": "getCardsByThemeForUser",
                "parameters": [
                    {
                        "name": "themeId",
                        "in": "path",
                        "description": "Identifiant du thème pour lequel récupérer les cartes",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Carte"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Accès non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la récupération des cartes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cartes": {
            "post": {
                "tags": [
                    "Cartes"
                ],
                "summary": "Crée une nouvelle carte pour un thème spécifique",
                "description": "Ajoute une nouvelle carte au thème spécifié, appartenant à l'utilisateur connecté.",
                "operationId": "createCard",
                "requestBody": {
                    "description": "Données de la nouvelle carte",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question",
                                    "reponse",
                                    "theme_id"
                                ],
                                "properties": {
                                    "question": {
                                        "description": "La question de la carte",
                                        "type": "string",
                                        "example": "Test question"
                                    },
                                    "reponse": {
                                        "description": "La réponse de la carte",
                                        "type": "string",
                                        "example": "Test réponse"
                                    },
                                    "theme_id": {
                                        "description": "Identifiant du thème auquel la carte sera ajoutée",
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 14
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carte créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carte"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données de la requête",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Détail des erreurs de validation",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la création de la carte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cartes/{carteId}": {
            "delete": {
                "tags": [
                    "Cartes"
                ],
                "summary": "Supprime une carte spécifique",
                "description": "Supprime une carte appartenant à l'utilisateur connecté, en vérifiant que la carte appartient bien à l'utilisateur.",
                "operationId": "deleteCard",
                "parameters": [
                    {
                        "name": "carteId",
                        "in": "path",
                        "description": "Identifiant de la carte à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carte supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Carte supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Erreur lors de la suppression de la carte ou accès non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la suppression de la carte ou accès non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Carte": {
                "title": "Carte",
                "description": "Schéma d'une carte associée à un thème",
                "required": [
                    "id",
                    "question",
                    "reponse",
                    "theme_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 66
                    },
                    "question": {
                        "type": "string",
                        "example": "Quelle est la définition de la dérivée d'une fonction ?"
                    },
                    "reponse": {
                        "type": "string",
                        "example": "Le taux de variation instantané"
                    },
                    "theme_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 14
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-22T17:57:30.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-22T17:57:30.000000Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}